streamlit_app.py
import os
import sys
import streamlit as st
import tempfile
import numpy as np
import cv2
from PIL import Image
import tensorflow as tf  # Import TensorFlow to ensure it's initialized

# Configure Streamlit
st.set_page_config(
    page_title="AI Face Swap",
    page_icon="🎭",
    layout="wide"
)

st.title("🎭 AI Face Swap")
st.caption("Upload a source face and a target image/video to swap faces")

# Simple file upload interface
source_file = st.file_uploader("Upload source face image", type=["jpg", "jpeg", "png"])
target_file = st.file_uploader("Upload target video/image", type=["jpg", "jpeg", "png", "mp4"])

col1, col2 = st.columns(2)

# Display uploaded files
if source_file is not None:
    with col1:
        st.image(source_file, caption="Source Face", use_column_width=True)
        
if target_file is not None:
    with col2:
        file_type = target_file.name.split(".")[-1].lower()
        if file_type in ["jpg", "jpeg", "png"]:
            st.image(target_file, caption="Target Image", use_column_width=True)
        elif file_type == "mp4":
            st.video(target_file, format="video/mp4")

# Processing settings
with st.sidebar:
    st.header("Settings")
    enhance_face = st.checkbox("Enhance face quality", value=True)
    keep_original_size = st.checkbox("Keep original size", value=True)
    many_faces = st.checkbox("Process multiple faces", value=False)

# Process button
if st.button("Swap Face", disabled=(source_file is None or target_file is None)):
    st.info("Processing... (this could take a while)")
    
    # Create a progress bar
    progress_bar = st.progress(0)
    
    # Sample placeholder code (doesn't actually do face swapping)
    with tempfile.TemporaryDirectory() as temp_dir:
        if source_file and target_file:
            # Update progress
            progress_bar.progress(10)
            
            # Save the uploaded files
            source_path = os.path.join(temp_dir, "source.jpg")
            with open(source_path, "wb") as f:
                f.write(source_file.getbuffer())
            
            # Update progress
            progress_bar.progress(30)
                
            target_path = os.path.join(temp_dir, "target." + target_file.name.split(".")[-1])
            with open(target_path, "wb") as f:
                f.write(target_file.getbuffer())
            
            # Update progress    
            progress_bar.progress(50)
            
            st.success(f"Files saved to temporary directory. Ready for processing.")
            
            # Try to import and use the roop library if possible
            try:
                # Add the current directory to path
                if os.path.dirname(os.path.abspath(__file__)) not in sys.path:
                    sys.path.append(os.path.dirname(os.path.abspath(__file__)))
                
                # Try to import roop functionality
                from roop.face_analyser import get_face_analyser
                from roop.utilities import is_image, is_video
                
                # Update progress
                progress_bar.progress(70)
                
                # Let the user know we found Roop modules
                st.success("Successfully imported Roop modules! Processing with actual face swap will be available soon.")
                
                # Future implementation will call actual roop functions:
                # result = roop.swap_face(source_path, target_path, enhance_face, many_faces)
                
            except ImportError as e:
                st.warning(f"Couldn't import Roop modules: {e}")
                st.info("Running in demo mode without actual face swapping.")
            
            # Update progress
            progress_bar.progress(90)
                
            # For now just display the original target as a placeholder
            st.subheader("Output Preview")
            if target_file.name.endswith(('.jpg', '.jpeg', '.png')):
                st.image(target_file, caption="Processed Result (Placeholder)", use_column_width=True)
            else:
                st.video(target_file, format="video/mp4")
                
            # Complete progress
            progress_bar.progress(100)

# Add system information in expander
with st.expander("System Information"):
    st.info(f"Python version: {sys.version}")
    st.info(f"OpenCV version: {cv2.__version__}")
    st.info(f"TensorFlow version: {tf.__version__}")
    st.info(f"NumPy version: {np.__version__}")
    
    # Check for GPU
    if tf.config.list_physical_devices('GPU'):
        st.success("TensorFlow is using GPU acceleration")
    else:
        st.warning("TensorFlow is using CPU only")

# Add footer
st.markdown("---")
st.caption("Created for educational purposes only.")
